research-collab-backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
python-multipart
cryptography
requests
